requirements.txt
Django==5.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==3.0.1
whitenoise==6.7.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
django-select2==8.4.0
weasyprint==56.0
pydyf==0.11.0
cairocffi==1.7.1
CairoSVG==2.8.2
djangorestframework==3.15.1

